Initial Commit
Hi there, this is my initial contribution to enRoute. In this commit I
have fixed the code to export gogo command properly. The aim of this
commit is to understand the rules for contributing to enRoute. Please
let me know if you have a guide if not please check my questions below.
My second objective is commit Unit test classes. However before I
proceed here are some questions:

1- do you have a formatting file for Eclipse, I do not wish to change
the sourcce formatting through commits?
2- can I add bnd.bndrun to create debug runtime?
3- can I modify files such as .gitignore?

Signed-off-by: daghan <daghanacay@gmail.com>